requirements.txt
# Core dependencies for Post-Discharge Medical AI Assistant POC

# LangChain and LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langgraph>=0.0.20

# OpenAI API
openai>=1.0.0

# Vector store and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0

# Web search
duckduckgo-search>=4.0.0

# Web framework
streamlit>=1.28.0

# Environment variables
python-dotenv>=1.0.0

# Data handling
pydantic>=2.0.0

# Text processing
pypdf>=3.0.0

# Utilities
typing-extensions>=4.5.0


